requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
tokenizers>=0.15.0

# Arabic NLP
farasa>=0.0.14

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
gdown>=4.7.1
rarfile>=4.0
PyYAML>=6.0

# Optional dependencies for specific features
# openpyxl>=3.1.0  # For AJGT dataset (Excel files)
# requests>=2.31.0  # For downloading datasets
# beautifulsoup4>=4.12.0  # For web scraping (if needed)

